requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
supabase==2.0.2
asyncpg==0.29.0
sqlalchemy==2.0.23

# Redis for caching and queues
redis==5.0.1
aioredis==2.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# AI and Image Processing
ultralytics==8.0.196
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
google-generativeai==0.3.2

# Geospatial calculations
geopy==2.4.0

# Background tasks
celery==5.3.4

# Environment and configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# HTTP client
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# File handling
aiofiles==23.2.0

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
